Move theme context into a separate folder
